api/requirements-test.txt
# Testing dependencies for LY-Merch API

# Core testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Test utilities
httpx>=0.24.0
requests>=2.28.0
faker>=18.0.0

# Database testing
pytest-postgresql>=4.1.0
sqlalchemy-utils>=0.41.0

# Performance and parallel testing
pytest-xdist>=3.0.0
pytest-benchmark>=4.0.0

# Coverage and reporting
coverage[toml]>=7.0.0

# API testing
fastapi[test]>=0.95.0
starlette[test]>=0.27.0

# Development utilities (optional)
pytest-sugar>=0.9.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0